# backend/internal/handlers/integracao_handler.go
package handlers

import (
	"bytes"
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"strconv"
	"time"

	"github.com/ceialmilk/api/internal/auth"
	"github.com/ceialmilk/api/internal/models"
	"github.com/ceialmilk/api/internal/response"
	"github.com/ceialmilk/api/internal/service"
	"github.com/gin-gonic/gin"
)

type IntegracaoHandler struct {
	integracaoSvc    *service.IntegracaoService
	animalSvc        *service.AnimalService
	toqueSvc         *service.DiagnosticoGestacaoService
	coberturaSvc     *service.CoberturaService
	animalSaudeSvc   *service.AnimalSaudeService
	alertaSvc        *service.AlertaService
}

func NewIntegracaoHandler(
	integracaoSvc *service.IntegracaoService,
	animalSvc *service.AnimalService,
	toqueSvc *service.DiagnosticoGestacaoService,
	coberturaSvc *service.CoberturaService,
	animalSaudeSvc *service.AnimalSaudeService,
	alertaSvc *service.AlertaService,
) *IntegracaoHandler {
	return &IntegracaoHandler{
		integracaoSvc:  integracaoSvc,
		animalSvc:      animalSvc,
		toqueSvc:       toqueSvc,
		coberturaSvc:   coberturaSvc,
		animalSaudeSvc: animalSaudeSvc,
		alertaSvc:      alertaSvc,
	}
}

func (h *IntegracaoHandler) Me(c *gin.Context) {
	clientID, _ := auth.GetIntegrationClientID(c)
	cliente, err := h.integracaoSvc.GetByID(c.Request.Context(), clientID)
	if err != nil {
		response.ErrorInternal(c, "Erro ao obter cliente", err.Error())
		return
	}
	resp := gin.H{
		"id":          cliente.ID,
		"nome":        cliente.Nome,
		"scopes":      cliente.Scopes,
		"fazenda_ids": cliente.FazendaIDs,
		"ativo":       cliente.Ativo,
	}
	response.SuccessOK(c, resp, "OK")
}

func (h *IntegracaoHandler) SearchAnimais(c *gin.Context) {
	identificacao := c.Query("identificacao")
	fazendaID, _ := strconv.ParseInt(c.Query("fazenda_id"), 10, 64)
	if identificacao == "" || fazendaID <= 0 {
		response.ErrorBadRequest(c, "identificacao e fazenda_id sao obrigatorios", nil)
		return
	}
	if !ValidateFazendaIntegracao(c, fazendaID) {
		return
	}
	list, err := h.animalSvc.SearchByIdentificacaoForFazendas(
		c.Request.Context(), identificacao, []int64{fazendaID}, true,
	)
	if err != nil {
		response.ErrorInternal(c, "Erro ao buscar", err.Error())
		return
	}
	response.SuccessOK(c, list, "OK")
}

func (h *IntegracaoHandler) GetAnimal(c *gin.Context) {
	id, _ := strconv.ParseInt(c.Param("id"), 10, 64)
	animal, err := h.animalSvc.GetByID(c.Request.Context(), id)
	if err != nil {
		response.ErrorNotFound(c, "Animal nao encontrado")
		return
	}
	if !ValidateFazendaIntegracao(c, animal.FazendaID) {
		return
	}
	response.SuccessOK(c, animal, "OK")
}

func (h *IntegracaoHandler) ListCoberturas(c *gin.Context) {
	animalID, _ := strconv.ParseInt(c.Query("animal_id"), 10, 64)
	if animalID <= 0 {
		response.ErrorBadRequest(c, "animal_id obrigatorio", nil)
		return
	}
	animal, err := h.animalSvc.GetByID(c.Request.Context(), animalID)
	if err != nil {
		response.ErrorNotFound(c, "Animal nao encontrado")
		return
	}
	if !ValidateFazendaIntegracao(c, animal.FazendaID) {
		return
	}
	list, err := h.coberturaSvc.GetByAnimalID(c.Request.Context(), animalID)
	if err != nil {
		response.ErrorInternal(c, "Erro ao listar coberturas", err.Error())
		return
	}
	response.SuccessOK(c, list, "OK")
}

func (h *IntegracaoHandler) CreateToque(c *gin.Context) {
	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		response.ErrorBadRequest(c, "body invalido", nil)
		return
	}
	c.Request.Body = io.NopCloser(bytes.NewReader(body))

	clientID, _ := auth.GetIntegrationClientID(c)
	idemKey := c.GetHeader("Idempotency-Key")
	reqHash := service.HashRequestBody(body)
	if cached, status, conflict, err := h.integracaoSvc.CheckIdempotency(c.Request.Context(), clientID, idemKey, reqHash); err != nil {
		response.ErrorInternal(c, "Erro de idempotencia", err.Error())
		return
	} else if conflict {
		response.Error(c, http.StatusConflict, response.CodeConflict, "Idempotency-Key ja usada com payload diferente", nil)
		return
	} else if cached != nil {
		var payload interface{}
		_ = json.Unmarshal(cached, &payload)
		response.Success(c, status, payload, "Resposta idempotente")
		return
	}

	var req struct {
		AnimalID                int64   `json:"animal_id" binding:"required"`
		Data                    string  `json:"data" binding:"required"`
		Resultado               string  `json:"resultado"`
		ClassificacaoOperacional *string `json:"classificacao_operacional"`
		FazendaID               int64   `json:"fazenda_id" binding:"required"`
		CoberturaID             *int64  `json:"cobertura_id"`
		DiasGestacaoEstimados   *int    `json:"dias_gestacao_estimados"`
		Metodo                  *string `json:"metodo"`
		Veterinario             *string `json:"veterinario"`
		Observacoes             *string `json:"observacoes"`
	}
	if err := json.Unmarshal(body, &req); err != nil {
		response.ErrorValidation(c, "Dados invalidos", err.Error())
		return
	}
	if !ValidateFazendaIntegracao(c, req.FazendaID) {
		return
	}
	t, err := time.Parse(time.RFC3339, req.Data)
	if err != nil {
		response.ErrorValidation(c, "data invalida", err.Error())
		return
	}
	d := &models.DiagnosticoGestacao{
		AnimalID: req.AnimalID, Data: t, Resultado: req.Resultado, FazendaID: req.FazendaID,
		ClassificacaoOperacional: req.ClassificacaoOperacional,
		CoberturaID: req.CoberturaID, DiasGestacaoEstimados: req.DiasGestacaoEstimados,
		Metodo: req.Metodo, Veterinario: req.Veterinario, Observacoes: req.Observacoes,
	}
	if actorID, ok := GetActorUserID(c); ok {
		d.CreatedBy = &actorID
	}
	if err := h.toqueSvc.Create(c.Request.Context(), d); err != nil {
		mapToqueError(c, err)
		return
	}
	wrap := response.SuccessResponse{Data: d, Message: "Toque registrado", Timestamp: time.Now().UTC().Format(time.RFC3339)}
	_ = h.integracaoSvc.SaveIdempotency(c.Request.Context(), clientID, idemKey, reqHash, http.StatusCreated, wrap)
	response.SuccessCreated(c, d, "Toque registrado")
}

func (h *IntegracaoHandler) CreateToqueLote(c *gin.Context) {
	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		response.ErrorBadRequest(c, "body invalido", nil)
		return
	}
	clientID, _ := auth.GetIntegrationClientID(c)
	idemKey := c.GetHeader("Idempotency-Key")
	if idemKey == "" {
		var bodyMap struct {
			IdempotencyKey *string `json:"idempotency_key"`
		}
		_ = json.Unmarshal(body, &bodyMap)
		if bodyMap.IdempotencyKey != nil {
			idemKey = *bodyMap.IdempotencyKey
		}
	}
	reqHash := service.HashRequestBody(body)
	if cached, status, conflict, err := h.integracaoSvc.CheckIdempotency(c.Request.Context(), clientID, idemKey, reqHash); err != nil {
		response.ErrorInternal(c, "Erro de idempotencia", err.Error())
		return
	} else if conflict {
		response.Error(c, http.StatusConflict, response.CodeConflict, "Idempotency-Key ja usada com payload diferente", nil)
		return
	} else if cached != nil {
		var payload interface{}
		_ = json.Unmarshal(cached, &payload)
		response.Success(c, status, payload, "Resposta idempotente")
		return
	}

	var req struct {
		FazendaID      int64                  `json:"fazenda_id" binding:"required"`
		IdempotencyKey *string                `json:"idempotency_key"`
		Itens          []models.ToqueLoteItem `json:"itens" binding:"required"`
	}
	if err := json.Unmarshal(body, &req); err != nil {
		response.ErrorValidation(c, "Dados invalidos", err.Error())
		return
	}
	if !ValidateFazendaIntegracao(c, req.FazendaID) {
		return
	}
	actorID, _ := GetActorUserID(c)
	loteSvc := service.NewIntegracaoToqueLoteService(h.animalSvc, h.toqueSvc, auth.GetIntegrationFazendaIDs(c), actorID)
	result, err := loteSvc.Process(c.Request.Context(), req.FazendaID, req.Itens)
	if err != nil {
		response.ErrorForbidden(c, err.Error())
		return
	}
	wrap := response.SuccessResponse{Data: result, Message: "Lote processado", Timestamp: time.Now().UTC().Format(time.RFC3339)}
	_ = h.integracaoSvc.SaveIdempotency(c.Request.Context(), clientID, idemKey, reqHash, http.StatusOK, wrap)
	response.SuccessOK(c, result, "Lote processado")
}

func (h *IntegracaoHandler) CreateCobertura(c *gin.Context) {
	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		response.ErrorBadRequest(c, "body invalido", nil)
		return
	}
	c.Request.Body = io.NopCloser(bytes.NewReader(body))

	clientID, _ := auth.GetIntegrationClientID(c)
	idemKey := c.GetHeader("Idempotency-Key")
	reqHash := service.HashRequestBody(body)
	if cached, status, conflict, err := h.integracaoSvc.CheckIdempotency(c.Request.Context(), clientID, idemKey, reqHash); err != nil {
		response.ErrorInternal(c, "Erro de idempotencia", err.Error())
		return
	} else if conflict {
		response.Error(c, http.StatusConflict, response.CodeConflict, "Idempotency-Key ja usada com payload diferente", nil)
		return
	} else if cached != nil {
		var payload interface{}
		_ = json.Unmarshal(cached, &payload)
		response.Success(c, status, payload, "Resposta idempotente")
		return
	}

	var req struct {
		AnimalID      int64   `json:"animal_id" binding:"required"`
		Tipo          string  `json:"tipo" binding:"required"`
		Data          string  `json:"data" binding:"required"`
		FazendaID     int64   `json:"fazenda_id" binding:"required"`
		CioID         *int64  `json:"cio_id"`
		TouroAnimalID *int64  `json:"touro_animal_id"`
		TouroInfo     *string `json:"touro_info"`
		SemenPartida  *string `json:"semen_partida"`
		Tecnico       *string `json:"tecnico"`
		ProtocoloID   *int64  `json:"protocolo_id"`
		Observacoes   *string `json:"observacoes"`
	}
	if err := json.Unmarshal(body, &req); err != nil {
		response.ErrorValidation(c, "Dados invalidos", err.Error())
		return
	}
	if !ValidateFazendaIntegracao(c, req.FazendaID) {
		return
	}
	t, err := time.Parse(time.RFC3339, req.Data)
	if err != nil {
		response.ErrorValidation(c, "data invalida", err.Error())
		return
	}
	cobertura := &models.Cobertura{
		AnimalID: req.AnimalID, Tipo: req.Tipo, Data: t, FazendaID: req.FazendaID,
		CioID: req.CioID, TouroAnimalID: req.TouroAnimalID, TouroInfo: req.TouroInfo,
		SemenPartida: req.SemenPartida, Tecnico: req.Tecnico, ProtocoloID: req.ProtocoloID, Observacoes: req.Observacoes,
	}
	if actorID, ok := GetActorUserID(c); ok {
		cobertura.CreatedBy = &actorID
	}
	if err := h.coberturaSvc.Create(c.Request.Context(), cobertura); err != nil {
		mapCoberturaError(c, err)
		return
	}
	wrap := response.SuccessResponse{Data: cobertura, Message: "Cobertura registrada", Timestamp: time.Now().UTC().Format(time.RFC3339)}
	_ = h.integracaoSvc.SaveIdempotency(c.Request.Context(), clientID, idemKey, reqHash, http.StatusCreated, wrap)
	response.SuccessCreated(c, cobertura, "Cobertura registrada")
}

func (h *IntegracaoHandler) CreateCoberturaLote(c *gin.Context) {
	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		response.ErrorBadRequest(c, "body invalido", nil)
		return
	}
	clientID, _ := auth.GetIntegrationClientID(c)
	idemKey := c.GetHeader("Idempotency-Key")
	if idemKey == "" {
		var bodyMap struct {
			IdempotencyKey *string `json:"idempotency_key"`
		}
		_ = json.Unmarshal(body, &bodyMap)
		if bodyMap.IdempotencyKey != nil {
			idemKey = *bodyMap.IdempotencyKey
		}
	}
	reqHash := service.HashRequestBody(body)
	if cached, status, conflict, err := h.integracaoSvc.CheckIdempotency(c.Request.Context(), clientID, idemKey, reqHash); err != nil {
		response.ErrorInternal(c, "Erro de idempotencia", err.Error())
		return
	} else if conflict {
		response.Error(c, http.StatusConflict, response.CodeConflict, "Idempotency-Key ja usada com payload diferente", nil)
		return
	} else if cached != nil {
		var payload interface{}
		_ = json.Unmarshal(cached, &payload)
		response.Success(c, status, payload, "Resposta idempotente")
		return
	}

	var req struct {
		FazendaID      int64                     `json:"fazenda_id" binding:"required"`
		IdempotencyKey *string                   `json:"idempotency_key"`
		Itens          []models.CoberturaLoteItem `json:"itens" binding:"required"`
	}
	if err := json.Unmarshal(body, &req); err != nil {
		response.ErrorValidation(c, "Dados invalidos", err.Error())
		return
	}
	if !ValidateFazendaIntegracao(c, req.FazendaID) {
		return
	}
	actorID, _ := GetActorUserID(c)
	loteSvc := service.NewIntegracaoCoberturaLoteService(h.animalSvc, h.coberturaSvc, auth.GetIntegrationFazendaIDs(c), actorID)
	result, err := loteSvc.Process(c.Request.Context(), req.FazendaID, req.Itens)
	if err != nil {
		response.ErrorForbidden(c, err.Error())
		return
	}
	wrap := response.SuccessResponse{Data: result, Message: "Lote processado", Timestamp: time.Now().UTC().Format(time.RFC3339)}
	_ = h.integracaoSvc.SaveIdempotency(c.Request.Context(), clientID, idemKey, reqHash, http.StatusOK, wrap)
	response.SuccessOK(c, result, "Lote processado")
}

func (h *IntegracaoHandler) ListSaude(c *gin.Context) {
	fazendaID, _ := strconv.ParseInt(c.Query("fazenda_id"), 10, 64)
	animalID, _ := strconv.ParseInt(c.Query("animal_id"), 10, 64)
	if fazendaID <= 0 || animalID <= 0 {
		response.ErrorBadRequest(c, "fazenda_id e animal_id sao obrigatorios", nil)
		return
	}
	if !ValidateFazendaIntegracao(c, fazendaID) {
		return
	}
	animal, err := h.animalSvc.GetByID(c.Request.Context(), animalID)
	if err != nil {
		response.ErrorNotFound(c, "Animal nao encontrado")
		return
	}
	if animal.FazendaID != fazendaID {
		response.ErrorForbidden(c, "Animal nao pertence a esta fazenda")
		return
	}
	list, err := h.animalSaudeSvc.ListByAnimalID(c.Request.Context(), animalID)
	if err != nil {
		mapSaudeError(c, err, "Erro ao listar casos de saude")
		return
	}
	if list == nil {
		list = []*models.AnimalSaude{}
	}
	response.SuccessOK(c, list, "OK")
}

func (h *IntegracaoHandler) CreateSaude(c *gin.Context) {
	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		response.ErrorBadRequest(c, "body invalido", nil)
		return
	}
	c.Request.Body = io.NopCloser(bytes.NewReader(body))

	clientID, _ := auth.GetIntegrationClientID(c)
	idemKey := c.GetHeader("Idempotency-Key")
	reqHash := service.HashRequestBody(body)
	if cached, status, conflict, err := h.integracaoSvc.CheckIdempotency(c.Request.Context(), clientID, idemKey, reqHash); err != nil {
		response.ErrorInternal(c, "Erro de idempotencia", err.Error())
		return
	} else if conflict {
		response.Error(c, http.StatusConflict, response.CodeConflict, "Idempotency-Key ja usada com payload diferente", nil)
		return
	} else if cached != nil {
		var payload interface{}
		_ = json.Unmarshal(cached, &payload)
		response.Success(c, status, payload, "Resposta idempotente")
		return
	}

	var req struct {
		AnimalID    int64   `json:"animal_id" binding:"required"`
		FazendaID   int64   `json:"fazenda_id" binding:"required"`
		TipoCaso    string  `json:"tipo_caso" binding:"required"`
		DataInicio  string  `json:"data_inicio" binding:"required"`
		DataFim     *string `json:"data_fim"`
		Status      string  `json:"status" binding:"required"`
		Observacoes *string `json:"observacoes"`
	}
	if err := json.Unmarshal(body, &req); err != nil {
		response.ErrorValidation(c, "Dados invalidos", err.Error())
		return
	}
	if !ValidateFazendaIntegracao(c, req.FazendaID) {
		return
	}
	animal, err := h.animalSvc.GetByID(c.Request.Context(), req.AnimalID)
	if err != nil {
		response.ErrorNotFound(c, "Animal nao encontrado")
		return
	}
	if animal.FazendaID != req.FazendaID {
		response.ErrorForbidden(c, "Animal nao pertence a esta fazenda")
		return
	}
	in, ok := parseSaveAnimalSaudeInput(c, saveAnimalSaudeRequest{
		TipoCaso: req.TipoCaso, DataInicio: req.DataInicio, DataFim: req.DataFim,
		Status: req.Status, Observacoes: req.Observacoes,
	})
	if !ok {
		return
	}
	if actorID, exists := GetActorUserID(c); exists {
		in.CreatedBy = &actorID
	}
	row, err := h.animalSaudeSvc.Create(c.Request.Context(), req.AnimalID, in)
	if err != nil {
		mapSaudeError(c, err, "Erro ao registrar caso de saude")
		return
	}
	wrap := response.SuccessResponse{Data: row, Message: "Caso de saude registrado", Timestamp: time.Now().UTC().Format(time.RFC3339)}
	_ = h.integracaoSvc.SaveIdempotency(c.Request.Context(), clientID, idemKey, reqHash, http.StatusCreated, wrap)
	response.SuccessCreated(c, row, "Caso de saude registrado")
}

func (h *IntegracaoHandler) ListAlertas(c *gin.Context) {
	fazendaID, _ := strconv.ParseInt(c.Query("fazenda_id"), 10, 64)
	if fazendaID <= 0 {
		response.ErrorBadRequest(c, "fazenda_id obrigatorio", nil)
		return
	}
	if !ValidateFazendaIntegracao(c, fazendaID) {
		return
	}
	limit := parseQueryIntPositiveDef(c.Query("limit"), 25)
	offset := parseQueryIntNonNeg(c.DefaultQuery("offset", "0"), 0)
	if limit > 100 {
		limit = 100
	}
	list, total, err := h.alertaSvc.ListByFazenda(c.Request.Context(), fazendaID, service.AlertaListQuery{
		Status:     c.Query("status"),
		Tipo:       c.Query("tipo"),
		Severidade: c.Query("severidade"),
		Limit:      limit,
		Offset:     offset,
	})
	if err != nil {
		mapAlertaIntegracaoError(c, err)
		return
	}
	response.SuccessOK(c, gin.H{"alertas": list, "total": total}, "OK")
}

func mapSaudeError(c *gin.Context, err error, internalMsg string) {
	if RespondIfDomainWriteError(c, err) {
		return
	}
	switch {
	case errors.Is(err, service.ErrAnimalNotFound):
		response.ErrorNotFound(c, "Animal nao encontrado")
	case errors.Is(err, service.ErrAnimalSaudeNotFound):
		response.ErrorNotFound(c, "Caso de saude nao encontrado")
	case errors.Is(err, service.ErrAnimalSaudeTipoCasoInvalido),
		errors.Is(err, service.ErrAnimalSaudeStatusInvalido),
		errors.Is(err, service.ErrAnimalSaudeDataFimInvalida):
		response.ErrorValidation(c, err.Error(), nil)
	default:
		response.ErrorInternal(c, internalMsg, err.Error())
	}
}

func mapAlertaIntegracaoError(c *gin.Context, err error) {
	switch {
	case errors.Is(err, service.ErrAlertaStatusInvalido),
		errors.Is(err, service.ErrAlertaTipoInvalido),
		errors.Is(err, service.ErrAlertaSeveridadeInvalida):
		response.ErrorValidation(c, err.Error(), nil)
	default:
		response.ErrorInternal(c, "Erro ao listar alertas", err.Error())
	}
}

func mapToqueError(c *gin.Context, err error) {
	if RespondIfIntegridadeCiclo(c, err) {
		return
	}
	if RespondIfAnimalForaRebanho(c, err) {
		return
	}
	switch {
	case errors.Is(err, service.ErrToquePositivoSemCobertura),
		errors.Is(err, service.ErrToquePositivoGestacaoAtiva),
		errors.Is(err, service.ErrResultadoOuClassificacaoObrigatorio),
		errors.Is(err, service.ErrClassificacaoOperacionalInvalida),
		errors.Is(err, service.ErrMetodoDiagnosticoInvalido),
		errors.Is(err, models.ErrClassificacaoResultadoInconsistente):
		response.ErrorValidation(c, err.Error(), nil)
	case err != nil && err.Error() == "resultado invalido":
		response.ErrorValidation(c, err.Error(), nil)
	default:
		response.ErrorInternal(c, "Erro ao registrar toque", err.Error())
	}
}

func mapCoberturaError(c *gin.Context, err error) {
	if RespondIfDomainWriteError(c, err) {
		return
	}
	switch {
	case errors.Is(err, service.ErrAnimalNotFound):
		response.ErrorNotFound(c, "Animal nao encontrado")
	case errors.Is(err, service.ErrCoberturaCamposObrigatorios),
		errors.Is(err, service.ErrCoberturaTipoInvalido),
		errors.Is(err, service.ErrCoberturaAnimalFazendaDiferente),
		errors.Is(err, service.ErrCoberturaApenasFemea),
		errors.Is(err, service.ErrCoberturaReprodutorObrigatorio),
		errors.Is(err, service.ErrCoberturaReprodutorNaoEncontrado),
		errors.Is(err, service.ErrCoberturaReprodutorInvalido):
		response.ErrorValidation(c, err.Error(), nil)
	default:
		response.ErrorInternal(c, "Erro ao registrar cobertura", err.Error())
	}
}