# backend/openapi/integracoes-v1.openapi.yaml
...
openapi: 3.0.3
info:
  title: CeialMilk API de Integrações
  description: |
    API máquina-a-máquina para sistemas externos e agentes de IA.
    Autenticação via `Authorization: Bearer cmk_live_...` (gerada em Admin → Integrações).

    **Importante:** `fazenda_id` e `identificacao` em buscas vão na **query string** da URL, não em headers.

    Documentação interativa: `/api/v1/integracoes/docs`
  version: 1.0.0
  contact:
    name: CeialMilk
servers:
  - url: /
    description: Servidor atual (mesmo host onde abriu o Swagger ou importou a spec)
...
security:
  - IntegrationApiKey: []
paths:
...
  /api/v1/integracoes/animais/search:
    get:
      tags: [Animais]
      summary: Buscar animais por identificação
      description: |
        Parâmetros obrigatórios na **query string** (não usar headers).
        Requer scope `animais:read` e fazenda vinculada ao cliente.
      operationId: searchAnimais
      parameters:
        - $ref: "#/components/parameters/FazendaIdQuery"
        - $ref: "#/components/parameters/IdentificacaoQuery"
      responses:
        "200":
          description: Lista de animais (pode ser vazia)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessAnimalList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/integracoes/animais/{id}:
    get:
      tags: [Animais]
      summary: Obter animal por ID
      operationId: getAnimal
      parameters:
        - $ref: "#/components/parameters/AnimalIdPath"
      responses:
        "200":
          description: Animal encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessAnimal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/integracoes/coberturas:
    get:
      tags: [Coberturas]
      summary: Listar coberturas do animal
      operationId: listCoberturas
      parameters:
        - name: animal_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID do animal
      responses:
        "200":
          description: Lista de coberturas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessCoberturaList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Coberturas]
      summary: Registar uma cobertura
      description: |
        Regista inseminação/monta na matriz. Regras de domínio (fêmea, reprodutor em monta natural, etc.) aplicam-se no servidor.
        Header opcional `Idempotency-Key` para evitar duplicatas em reenvio.
      operationId: createCobertura
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCoberturaRequest"
      responses:
        "201":
          description: Cobertura criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessCobertura"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/integracoes/coberturas/lote:
    post:
      tags: [Coberturas]
      summary: Registar coberturas em lote
      description: |
        Processa cada item independentemente (sucesso parcial). Matriz identificada por `identificacao` (brinco).
        `Idempotency-Key` no header ou `idempotency_key` no body.
      operationId: createCoberturaLote
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoberturaLoteRequest"
      responses:
        "200":
          description: Lote processado (ver sucesso e falhas no body)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessCoberturaLote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/integracoes/toques:
    post:
      tags: [Toques]
      summary: Registar um toque
      description: |
        Cria diagnóstico de gestação. Toque POSITIVO exige cobertura (informada ou resolvida automaticamente).
        Header opcional `Idempotency-Key` para evitar duplicatas em reenvio.
      operationId: createToque
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToqueRequest"
      responses:
        "201":
          description: Toque criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessDiagnostico"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/integracoes/toques/lote:
    post:
      tags: [Toques]
      summary: Registar toques em lote
      description: |
        Processa cada item independentemente (sucesso parcial).
        `Idempotency-Key` no header ou `idempotency_key` no body.
      operationId: createToqueLote
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToqueLoteRequest"
      responses:
        "200":
          description: Lote processado (ver sucesso e falhas no body)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessToqueLote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/integracoes/saude:
    get:
      tags: [Saude]
      summary: Listar casos de saúde do animal
      description: |
        Requer scope `saude:read`, fazenda vinculada ao cliente e animal no rebanho.
        Parâmetros `fazenda_id` e `animal_id` obrigatórios na query string.
      operationId: listSaude
      parameters:
        - $ref: "#/components/parameters/FazendaIdQuery"
        - name: animal_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: ID do animal
      responses:
        "200":
          description: Lista de casos de saúde
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessAnimalSaudeList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Saude]
      summary: Registar caso de saúde
      description: |
        Regista resultado clínico/laboratorial para um animal. `created_by` é o actor da integração (BR-INTEG-001).
        Header opcional `Idempotency-Key` para evitar duplicatas em reenvio.
      operationId: createSaude
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegracaoCreateSaudeRequest"
      responses:
        "201":
          description: Caso de saúde criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessAnimalSaude"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/integracoes/alertas:
    get:
      tags: [Alertas]
      summary: Listar alertas da fazenda
      description: |
        Requer scope `alertas:read` e fazenda vinculada ao cliente.
        Filtros opcionais por status, tipo e severidade.
      operationId: listAlertas
      parameters:
        - $ref: "#/components/parameters/FazendaIdQuery"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/StatusAlerta"
        - name: tipo
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TipoAlerta"
        - name: severidade
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SeveridadeAlerta"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 25
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Lista paginada de alertas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessAlertasList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    IntegrationApiKey:
      type: http
      scheme: bearer
      bearerFormat: cmk_live
      description: |
        Chave API gerada em Admin → Integrações. Formato: `cmk_live_<segredo>`.
        No Swagger UI use Authorize e cole o valor completo (com prefixo cmk_live_).
  parameters:
    FazendaIdQuery:
      name: fazenda_id
      in: query
      required: true
      schema:
        type: integer
        format: int64
      description: ID da fazenda (deve estar vinculada ao cliente)
    IdentificacaoQuery:
      name: identificacao
      in: query
      required: true
      schema:
        type: string
      description: Brinco ou identificação do animal
    AnimalIdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: Chave de idempotência (reenvio com mesmo body devolve resposta guardada)
  responses:
    BadRequest:
      description: Pedido inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Chave inválida, revogada ou ausente
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Scope ou fazenda não autorizada
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Idempotency-Key já usada com payload diferente
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit excedido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    SuccessAnimal:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Animal"
    SuccessAnimalList:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Animal"
...
    CreateCoberturaRequest:
      type: object
      required: [animal_id, tipo, data, fazenda_id]
      properties:
        animal_id:
          type: integer
          format: int64
        tipo:
          $ref: "#/components/schemas/TipoCobertura"
        data:
          type: string
          format: date-time
          description: RFC3339, ex. 2026-05-20T10:00:00Z
        fazenda_id:
          type: integer
          format: int64
        cio_id:
          type: integer
          format: int64
          nullable: true
        touro_animal_id:
          type: integer
          format: int64
          nullable: true
        touro_info:
          type: string
          nullable: true
        semen_partida:
          type: string
          nullable: true
        tecnico:
          type: string
          nullable: true
        protocolo_id:
          type: integer
          format: int64
          nullable: true
        observacoes:
          type: string
          nullable: true
    SuccessCobertura:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Cobertura"
    SuccessCoberturaList:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Cobertura"
...
    CoberturaLoteRequest:
      type: object
      required: [fazenda_id, itens]
      properties:
        fazenda_id:
          type: integer
          format: int64
        idempotency_key:
          type: string
          maxLength: 128
        itens:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CoberturaLoteItem"
...
    SuccessCoberturaLote:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CoberturaLoteResult"
...
    SuccessDiagnostico:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/DiagnosticoGestacao"
...
    StatusAlerta:
      type: string
      enum: [ABERTO, EM_ANDAMENTO, RESOLVIDO, IGNORADO]
...